+++ code-gen/tests/py_gen_test.py
@@ -503,6 +503,12 @@ dictionary Foo { ///< inline comment
     float bar;
 };
 
+dictionary Bar {
+    /// Different comment for bar
+    float bar;
+    float foo; ///< inline comment for foo
+};
+
 /// Typedef comment
 typedef int my_int; ///< inline typedef comment
 
@@ -523,7 +529,7 @@ enum MyEnum {
     pattern = re.compile(r'"""\s*(.*?)\s*"""', re.DOTALL)
     matches = pattern.findall(result)
 
-    assert len(matches) == 10
+    assert len(matches) == 12
     assert "Typedef comment\n\ninline typedef comment" in matches[0]
     assert "My Enum comment" in matches[1]
     assert "Enum value 1 comment" in matches[2]
@@ -537,8 +543,10 @@ enum MyEnum {
     assert "inline comment" in matches[5]
     assert "Short comment for foo" in matches[6]
     assert "Short comment for bar" in matches[7]
-    assert "Load" in matches[8]
-    assert "Save" in matches[9]
+    assert "Different comment for bar" in matches[8]
+    assert "inline comment for foo" in matches[9]
+    assert "Load" in matches[10]
+    assert "Save" in matches[11]
 
 
 def test__render_template_string_default_value() -> None: